+++ Cassini/src/org/insa/xml/osm/OsmNodeRef.java
@@ -27,13 +27,13 @@ import org.simpleframework.xml.Root;
 public class OsmNodeRef {
     
     @Attribute
-    private int ref;
+    private long ref;
     
-    public int getRef() {
+    public long getRef() {
         return ref;
     }
     
-    public void setRef(int ref) {
+    public void setRef(long ref) {
         this.ref = ref;
     }
